python-fitness-trainer/requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
mediapipe==0.10.32
PyYAML==6.0.1
numpy<2.0.0
pyttsx3==2.90
google-genai==0.8.0